OpenCV FaceDetectorYN. Model
